Galloper & 'Active Directory group' integration issue
In production we faced with next situation -> Carrier KeyCloak integrated with customer's Active Directory, and users in AD has different internal 'groups' assigned to them. Almost all of them are not 'numeric' or 'superadmin'.
As the result, we are trying to pass into 'Int' filter 'non-int' values
SQL representation is:

'SELECT * FROM projects WHERE id IN (1, 'testGroupName')

At this point SQL driver will raise an exception and on the UI we will see '500' status code.